refactor(plugins): move loading to config
