+++ lib/config.js
@@ -14,7 +14,10 @@ module.exports = function (options) {
 
   var config = {
     name: pkg.name,
-    plugins: _.get(pkg, 'hoodie.plugins') || [],
+    plugins: require('./plugins/load')(
+      _.get(pkg, 'hoodie.plugins') || [],
+      projectPath
+    ),
     paths: {
       data: options.data || path.join(projectPath, 'data'),
       project: projectPath,
