+++ lib/index.js
@@ -29,14 +29,12 @@ exports.init = function (env_config, callback) {
   var hapi = require('hapi')
 
   var server = new hapi.Server(options)
-  var plugins = require('./plugins')
 
   async.applyEachSeries([
     require('./database/start'),
     require('./database/install'),
-    plugins.load,
     require('./bundle'),
-    plugins.startAll,
+    require('./plugins'),
     async.apply(exports.configureServer, server)
   ], env_config, function (err) {
     if (err) return callback(err)
