+++ lib/plugins/index.js
@@ -1,50 +1,14 @@
 var EventEmitter = require('events').EventEmitter
-var fs = require('fs')
 var path = require('path')
 var url = require('url')
 
-var _ = require('lodash')
 var async = require('async')
 var log = require('npmlog')
 
+var loadModule = require('./load-module')
 var plugins_manager = require('./manager')
 
-exports.load = function (env_config, callback) {
-  log.silly('plugins', 'Reading project\'s package.json for plugins')
-  if (!env_config.plugins.length) {
-    log.info('plugins', 'Nothing found in package.json')
-    return callback(null, env_config)
-  }
-
-  env_config.plugins = _(env_config.plugins)
-  .map(function (plugin) {
-    var pluginPath = exports.resolvePluginPath(env_config.paths.project, plugin)
-    var pkg = JSON.parse(fs.readFileSync(path.join(pluginPath, 'package.json')))
-    var hapiHook = exports.loadModule(path.join(pluginPath, 'hooks', 'hapi'))
-    var staticHook = exports.loadModule(path.join(pluginPath, 'hooks', 'static'))
-
-    return {
-      path: pluginPath,
-      pkg: pkg,
-      hooks: {
-        static: staticHook,
-        hapi: hapiHook ? _.defaultsDeep(hapiHook, {
-          attributes: {
-            name: pkg.name
-          }
-        }) : null
-      }
-    }
-  })
-  .mapKeys(function (plugin) {
-    return plugin.pkg.name
-  })
-  .value()
-
-  callback(null, env_config)
-}
-
-exports.startAll = function (env_config, callback) {
+var exports = module.exports = function (env_config, callback) {
   log.verbose('plugins', 'Starting all plugins')
 
   plugins_manager.start(url.format(env_config.db), function (err, manager) {
@@ -54,6 +18,7 @@ exports.startAll = function (env_config, callback) {
 
     // loop through plugins and start
     var plugins = Object.keys(env_config.plugins)
+
     async.map(plugins, function (name, cb) {
       var hoodie = manager.createAPI({name: name})
       hoodie._name = name
@@ -74,8 +39,13 @@ exports.startPlugin = function (name, env_config, hoodie, callback) {
 
   hoodie.env = env_config
 
-  var doStartPlugin = function (p, cb) {
-    var serverModule = exports.loadModule(path.join(p, 'server'))
+  async.applyEachSeries([
+    doStartPlugin,
+    doInitHooks
+  ], env_config.plugins[name].path, callback)
+
+  function doStartPlugin (pluginPath, cb) {
+    var serverModule = loadModule(path.join(pluginPath, 'server'))
     if (!serverModule) {
       log.silly('plugins', 'No server module found for ' + name)
       return cb()
@@ -84,8 +54,8 @@ exports.startPlugin = function (name, env_config, hoodie, callback) {
     return serverModule(hoodie, cb)
   }
 
-  var doInitHooks = function (p, cb) {
-    var dynamicHook = exports.loadModule(path.join(p, 'hooks', 'dynamic'))
+  function doInitHooks (pluginPath, cb) {
+    var dynamicHook = loadModule(path.join(pluginPath, 'hooks', 'dynamic'))
     if (!dynamicHook) {
       log.silly('plugins', 'No dynamic hook found for ' + name)
       return cb()
@@ -94,61 +64,4 @@ exports.startPlugin = function (name, env_config, hoodie, callback) {
     env_config.plugins[name].hooks.dynamic = hoodie.hooks
     dynamicHook(hoodie, cb)
   }
-
-  async.applyEachSeries([
-    doStartPlugin,
-    doInitHooks
-  ], env_config.plugins[name].path, callback)
-}
-
-exports.loadModule = function (p) {
-  try {
-    var mod = require(p)
-    return typeof mod === 'function' ? mod : null
-  } catch (e) {
-    if (e.code !== 'MODULE_NOT_FOUND') {
-      throw e
-    }
-    return null
-  }
-}
-
-exports.resolvePluginPath = function (projectDir, plugin) {
-  var errorMessage = 'Plugin not found: ' + plugin
-
-  // module lookup
-  if (plugin[0] !== '.' && plugin[0] !== '/') {
-    try {
-      return path.dirname(require.resolve(path.join(plugin, 'package.json')))
-    } catch (e) {
-      if (e.code !== 'MODULE_NOT_FOUND') {
-        throw new Error(errorMessage)
-      }
-
-      // hoodie-server might be `npm link`ed. Let's retry in the cwd.
-      try {
-        return path.dirname(require.resolve(path.join(
-          projectDir,
-          'node_modules',
-          plugin,
-          'package.json'
-        )))
-      } catch (e) {
-        throw new Error(errorMessage)
-      }
-    }
-  }
-
-  // relative or absolute path
-  var dir = path.resolve(projectDir, plugin)
-  try {
-    if (fs.statSync(dir).isDirectory()) {
-      return dir
-    }
-  } catch (e) {
-    if (e.code !== 'ENOENT') {
-      throw e
-    }
-  }
-  throw new Error(errorMessage)
 }
